+++ practitioner-service/src/test/java/uk/ac/ox/ndph/mts/practitioner_service/validation/RoleAssignmentValidationTests.java
@@ -12,7 +12,7 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.ac.ox.ndph.mts.practitioner_service.NullableConverter;
-import uk.ac.ox.ndph.mts.practitioner_service.client.RoleServiceClient;
+import uk.ac.ox.ndph.mts.roleserviceclient.RoleServiceClient;
 import uk.ac.ox.ndph.mts.practitioner_service.client.SiteServiceClient;
 import uk.ac.ox.ndph.mts.practitioner_service.exception.RestException;
 import uk.ac.ox.ndph.mts.practitioner_service.model.RoleAssignment;
